Pull section content out into separate view
